pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otobo_znuny"
version = "1.3.0"
description = "Python client library for the OTOBO / Znuny REST API"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Tobias Bück", email = "tab@softoft.de" }]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: AsyncIO"
]
dependencies = [
    "httpx>=0.27,<0.29",
    "pydantic>=2.11.3,<3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.28,<3",
    "typer>=0.13.1,<0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.18.2",
    "ruff>=0.13.1",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]

[project.urls]
homepage = "https://otobo-docs.softoft.de/administration/automation/rest-api"
documentation = "https://otobo-docs.softoft.de/administration/automation/rest-api"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.18.2",
    "ruff>=0.13.1",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]


[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 100
fix = true
unsafe-fixes = true
show-fixes = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = [
  # Core pycodestyle/pyflakes
  "E", "F", "W",
  # isort (import order)
  "I",
  # flake8-bugbear (common bugs & smells)
  "B",
  # pyupgrade (modernize for target-version)
  "UP",
  # pep8-naming
  "N",
  # import conventions
  "TID",
  # complexity & simplifications
  "C90", "C4", "RET", "SIM",
  # quality/safety
  "ARG", "BLE", "COM", "ISC", "PIE", "Q", "DTZ", "YTT",
  # tests/style
  "PT",
  # print / debug
  "T20",
  # Ruff-specific
  "RUF",
  # Security (Bandit rules via Ruff) — strict; tune if noisy
  "S",
]

[tool.ruff.lint.isort]
known-first-party = ["otobo_znuny"]
combine-as-imports = true
force-sort-within-sections = true


[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = ["src"]
namespace_packages = true
explicit_package_bases = true

warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true
warn_return_any = true
no_implicit_reexport = true

disallow_any_generics = true
disallow_subclassing_any = true

plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[mypy-mariadb]
ignore_missing_imports = true

[tool.tox]
requires = ["tox>=4", "tox-min-req>=0.0.9"]
env_list = ["py{311,312,313}-{latest,min}"]

[tool.tox.testenv]
package = "editable"
extras = []
deps = [
    "pytest",
    "pytest-asyncio"
]
commands = [["pytest", "-q"]]

[tool.tox.testenv.min_req]
min = 1